TODO
Interesting thing found. In debug build a type defined inside a .cpp
file that has the same name as another type in another .cpp file
produces the same component ID while this is not happening in release
build. Need to document this later. Not a real problem anyway, since
components are going to be called differently and defined through header
files where it won't be possible to have another type with the same name
in the same namespace.

Signed-off-by: Blyschak <stepanblischak@gmail.com>