Optimize chunk iterator comparison
The default generated equality operator checks equality of all pointers of a chunk iterator.
This is not required as the invariant is that all pointers have same offset from the start.
Therefore, it is enough to compare only the first pointer, assuming all the rest must be equal.

This shows up as a 2x improvement when iterating a large chunk doing very little work inside the loop.

Signed-off-by: Blyschak <stepanblischak@gmail.com>